+++ src/core/ref.jl
@@ -8,7 +8,7 @@ end
 
 
 "compute bus pair level data, can be run on data or ref data structures"
-function calc_buspair_parameters(buses, branches, conductor_ids, ismulticondcutor)
+function calc_buspair_parameters(buses, branches)
     bus_lookup = Dict(bus["index"] => bus for (i,bus) in buses if bus["bus_type"] != 4)
 
     branch_lookup = Dict(branch["index"] => branch for (i,branch) in branches if branch["br_status"] == 1 && haskey(bus_lookup, branch["f_bus"]) && haskey(bus_lookup, branch["t_bus"]))
@@ -17,28 +17,15 @@ function calc_buspair_parameters(buses, branches, conductor_ids, ismulticondcuto
 
     bp_branch = Dict((bp, typemax(Int)) for bp in buspair_indexes)
 
-    if ismulticondcutor
-        bp_angmin = Dict((bp, [-Inf for c in conductor_ids]) for bp in buspair_indexes)
-        bp_angmax = Dict((bp, [ Inf for c in conductor_ids]) for bp in buspair_indexes)
-    else
-        @assert(length(conductor_ids) == 1)
-        bp_angmin = Dict((bp, -Inf) for bp in buspair_indexes)
-        bp_angmax = Dict((bp,  Inf) for bp in buspair_indexes)
-    end
+    bp_angmin = Dict((bp, -Inf) for bp in buspair_indexes)
+    bp_angmax = Dict((bp,  Inf) for bp in buspair_indexes)
 
     for (l,branch) in branch_lookup
         i = branch["f_bus"]
         j = branch["t_bus"]
 
-        if ismulticondcutor
-            for c in conductor_ids
-                bp_angmin[(i,j)][c] = max(bp_angmin[(i,j)][c], branch["angmin"][c])
-                bp_angmax[(i,j)][c] = min(bp_angmax[(i,j)][c], branch["angmax"][c])
-            end
-        else
-            bp_angmin[(i,j)] = max(bp_angmin[(i,j)], branch["angmin"])
-            bp_angmax[(i,j)] = min(bp_angmax[(i,j)], branch["angmax"])
-        end
+        bp_angmin[(i,j)] = max(bp_angmin[(i,j)], branch["angmin"])
+        bp_angmax[(i,j)] = min(bp_angmax[(i,j)], branch["angmax"])
 
         bp_branch[(i,j)] = min(bp_branch[(i,j)], l)
     end